Hi, your friendly numist here again...
another artifact of the upgrade is that we lost magic_quotes, the most obvious (to users) effect of which was that some of their comments would get eaten.  on our side, that meant that none of our input was escaped and we were pretty unprotected against SQL injection attacks.  

with tvenhola at a browser and a vim window for myself, I've located and added escaping code to every incoming string that goes in a sql query.  I'll be revising this with my other changes while I'm on vacation.

git-svn-id: http://abadthing.com/tmbo/trunk@127 7ced0b6d-2a13-0410-a4a2-c2825b4e6d43