your friendly numist here, with a pre-big-upgrade change.
it looks like jerry added himself to the list for failed login attempt mailings

someone changed includes/textlinks.txt and includes/header.txt to link to hurgames dot com, which is something we shouldnt be doing with links that are supposed to point into tmbo.

my change adds a code path for upgrades so the site can't be used by users mid-upgrade, and admins are warned as they test the site/browse unawares.

git-svn-id: http://abadthing.com/tmbo/trunk@131 7ced0b6d-2a13-0410-a4a2-c2825b4e6d43