this change introduces some cleanup as to how posses and lists are drawn:
posses:
now that folks that joined before the invitation system we have now was created are marked as having referred themselves (as inaccurate as that may be in truth), parts of the site that deal with posses need to be aware that a person cannot be in their own posse - they own the posse.

lists:
in a lot of places (everywhere except discussions), extra-long lines were handled in an odd way.  this is still not perfect, but is a lot better now.  I think I got them all, but here are the views that have been modified:
* main
* hof
* user
* discussions

if you find any areas where extra-long lines are exiting their containers, please report.


git-svn-id: http://abadthing.com/tmbo/trunk@285 7ced0b6d-2a13-0410-a4a2-c2825b4e6d43