this change introduces a split between the frontend and backend of the site.  like a wedge, this change starts only with introducing the split, with OO representations of users, comments, and uploads, and splits only as far as comment threads.
the benefits?  comments.inc now makes one SQL call of its own, and the refactoring isn't even done yet.

also changing are the helper functions that you usually find in the top of each file.  the last checkin changed a lot of files, all of which had duplicate functionality.  now all thread subscription and management can be done using assets/comments.inc (although this might get rolled into the OO katamari as time goes on).

unifying a backend to run the site will help a lot when it comes to new features coming our way, including making the API and the site share the same backend code, which will simplify changes, especially as it becomes time to start restructuring parts of the database to fit new features on their way.


git-svn-id: http://abadthing.com/tmbo/trunk@293 7ced0b6d-2a13-0410-a4a2-c2825b4e6d43