let's try this out...
I'm getting tired of decorations on all our links.  if it's not black it's usually clickable, especially on the pictures page.

nsfw and tmbo tags on images are clickable, to toggle filtering.  if you don't figure that out quickly enough, the block at the bottom of pic.php is still there for you to set to your own liking for now.

it turns out that Good UI is not an easy problem, and that overall people aren't big fans of big changes.  we'll figure it out.


git-svn-id: http://abadthing.com/tmbo/trunk@318 7ced0b6d-2a13-0410-a4a2-c2825b4e6d43