this change implements search with Lucene.
unfortunately, this pulls in the Zend framework, which is a lot like adding boost to a project in that there are a ton of files.  at some point I plan on thinning us down to a lowest common subset, but I don't have the time to follow all the dependencies that spiderweb around the framework.

anyway.

searching works.  update to this change and all new uploads and comments will be indexed.  there's a script in the works that will add your old comments and files to the index as well.

this is going to be rolled out on the sandbox for a while before merging to the main site.

this change also fixes some longstanding issues with quickcomment, some rendering errors, and quite a bit of code cleanup.  among other notables, there is only one code path for posting a comment now.


git-svn-id: http://abadthing.com/tmbo/trunk@347 7ced0b6d-2a13-0410-a4a2-c2825b4e6d43