migration script for indexing existing comments and uploads
small fix for search that increases the relevance of results

the sandbox is chugging away and should be ready for in-depth testing by EOD tomorrow.  maybe earlier.  once tested, this will roll out to the main site and the index will be updated with the last day or two's comments and posts.


git-svn-id: http://abadthing.com/tmbo/trunk@348 7ced0b6d-2a13-0410-a4a2-c2825b4e6d43