checked and applied cor's patch.  quick comment box now remembers your comment when you accidentally (or not) close the qc box.
note that the text will be lost if you navigate away from the page and rely on your browser to save the content.  thinking client-side database storage might be a good solution for this if there's no elegant way of fixing this.  also attractive because it's a new toy to learn.

git-svn-id: http://abadthing.com/tmbo/trunk@382 7ced0b6d-2a13-0410-a4a2-c2825b4e6d43