support tokens as a method to log in
this change updates all users of `mustLogIn()` that pass an argument,
all users of `loggedin()` (now just `login()`) and all users of
`login()` that pass an argument.

this change includes a considerable amount of refactoring of the
authentication code. session, cookie, u/p, and token methods have all
been tested to work when logging in as an admin.

mixed in with this change are more changes to `me()` and places in the
code that should be calling it.