removing most use of $_SESSION in code
the $_SESSION variable should only be used by code that logs you in,
logs you off, or caches data (like postdata) between pages. under no
circumstances should normal code access `userid`, `username`, or
`status` in $_SESSION. use me() for that information, including
checking if there is a logged-in user.

this change also cleans up one or two queries, removes the old
dailyIndex, and simplifies the arguments to some functions that really
shouldn't be used on a user other than the one that's currently
logged-in anyway.