Edge case bugfix.
It's possible to initialize an Upload object with only an id and then
call getComments without incurring a call to Upload::getUploadRow
(either Upload::exists(), which would return bool, or another accessor,
which would trigger an error). We should be failing fast as possible,
so if the code didn't call exists() first, it should be willing to
accept a fast and hard error.